refactor: API 요청 처리를 Datasource에서 처리하도록 수정

+++ app/src/main/java/woowacourse/shopping/data/service/ProductService.kt
@@ -17,6 +17,6 @@ interface ProductService {
 
     @GET("/products/{id}")
     fun fetchProduct(
-        @Path("id") productId: Int,
+        @Path("id") productId: Long,
     ): Call<ProductResponse>
 }
